pyproject.toml
[project]
name = "mb-template"
version = "0.20260323.2"
requires-python = ">=3.13"
dependencies = [
    "pyserial>=3.5",
    "opencv-python-headless>=4.9",
    "mcp>=1.0",
    "numpy>=1.24",
    "textual>=0.80",
]

[project.scripts]
rrctl = "radio_relay.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["radio_relay"]
